+++ components/src/components/elements/checkout-dialog/Checkout.tsx
@@ -1,0 +1,114 @@
+import { useTheme } from "styled-components";
+import { type Stripe } from "@stripe/stripe-js";
+import { Elements } from "@stripe/react-stripe-js";
+import type {
+  CompanyPlanDetailResponseData,
+  SetupIntentResponseData,
+} from "../../../api";
+import { useEmbed } from "../../../hooks";
+import { PaymentForm, PaymentMethod } from "../../elements";
+import { Box, Text } from "../../ui";
+
+interface CheckoutProps {
+  period: string;
+  setPaymentMethodId: React.Dispatch<React.SetStateAction<string | undefined>>;
+  setShowPaymentForm: React.Dispatch<React.SetStateAction<boolean>>;
+  showPaymentForm: boolean;
+  stripe: Promise<Stripe | null> | null;
+  plan?: CompanyPlanDetailResponseData;
+  setupIntent?: SetupIntentResponseData;
+}
+
+export const Checkout = ({
+  period,
+  plan,
+  setPaymentMethodId,
+  setShowPaymentForm,
+  setupIntent,
+  showPaymentForm,
+  stripe,
+}: CheckoutProps) => {
+  const theme = useTheme();
+
+  const { data } = useEmbed();
+
+  return (
+    <>
+      {showPaymentForm && setupIntent?.setupIntentClientSecret ? (
+        <Elements
+          stripe={stripe}
+          options={{
+            appearance: {
+              theme: "stripe",
+              variables: {
+                // Base
+                fontFamily: '"Public Sans", system-ui, sans-serif',
+                spacingUnit: "0.25rem",
+                borderRadius: "0.5rem",
+                colorText: "#30313D",
+                colorBackground: "#FFFFFF",
+                colorPrimary: "#0570DE",
+                colorDanger: "#DF1B41",
+
+                // Layout
+                gridRowSpacing: "1.5rem",
+                gridColumnSpacing: "1.5rem",
+              },
+              rules: {
+                ".Label": {
+                  fontSize: "1rem",
+                  fontWeight: "400",
+                  marginBottom: "0.75rem",
+                  color: theme.typography.text.color,
+                },
+              },
+            },
+            clientSecret: setupIntent.setupIntentClientSecret,
+          }}
+        >
+          <Box $width="100%" $marginBottom="1.5rem">
+            <Text $size={18}>Add payment method</Text>
+          </Box>
+
+          <PaymentForm
+            plan={plan}
+            period={period}
+            onConfirm={(value) => {
+              setPaymentMethodId(value);
+            }}
+          />
+
+          {data.subscription?.paymentMethod && (
+            <Box>
+              <Text
+                onClick={() => setShowPaymentForm(false)}
+                $font={theme.typography.link.fontFamily}
+                $size={theme.typography.link.fontSize}
+                $weight={theme.typography.link.fontWeight}
+                $color={theme.typography.link.color}
+              >
+                Use existing payment method
+              </Text>
+            </Box>
+          )}
+        </Elements>
+      ) : (
+        <>
+          <PaymentMethod />
+
+          <Box>
+            <Text
+              onClick={() => setShowPaymentForm(true)}
+              $font={theme.typography.link.fontFamily}
+              $size={theme.typography.link.fontSize}
+              $weight={theme.typography.link.fontWeight}
+              $color={theme.typography.link.color}
+            >
+              Change payment method
+            </Text>
+          </Box>
+        </>
+      )}
+    </>
+  );
+};
